requirements.txt
djangorestframework
SimpleJWT
django
django-cors-headers
pandas==2.0.3
openpyxl==3.1.2